+++ src/Exceptions/ModelSettingsException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Glorand\Model\Settings\Exceptions;
+
+use Exception;
+
+class ModelSettingsException extends Exception
+{
+}