+++ src/Managers/TableSettingsManager.php
@@ -5,11 +5,20 @@ namespace Glorand\Model\Settings\Managers;
 use Glorand\Model\Settings\Contracts\SettingsManagerContract;
 use Glorand\Model\Settings\Models\ModelSettings;
 
+/**
+ * Class TableSettingsManager
+ * @package Glorand\Model\Settings\Managers
+ * @property  \Illuminate\Database\Eloquent\Model|\Glorand\Model\Settings\Traits\HasSettingsTable $model
+ */
 class TableSettingsManager extends AbstractSettingsManager
 {
+    /**
+     * @param array $settings
+     * @return \Glorand\Model\Settings\Contracts\SettingsManagerContract
+     */
     public function apply(array $settings = []): SettingsManagerContract
     {
-        if (!$modelSettings = $this->model->modelSettings) {
+        if (!$modelSettings = $this->model->modelSettings()->first()) {
             $modelSettings = new ModelSettings();
             $modelSettings->model()->associate($this->model);
         }
@@ -19,11 +28,17 @@ class TableSettingsManager extends AbstractSettingsManager
         return $this;
     }
 
+    /**
+     * @param string|null $path
+     * @return \Glorand\Model\Settings\Contracts\SettingsManagerContract
+     * @throws \Exception
+     */
     public function delete(string $path = null): SettingsManagerContract
     {
         if (!$path) {
-            if ($this->model->modelSettings) {
-                $this->model->modelSettings->delete();
+            /** @var ModelSettings $modelSettings */
+            if ($modelSettings = $this->model->modelSettings()->first()) {
+                $modelSettings->delete();
             }
         } else {
             $settings = $this->all();
@@ -34,6 +49,11 @@ class TableSettingsManager extends AbstractSettingsManager
         return $this;
     }
 
+    /**
+     * @param string $path
+     * @param mixed $value
+     * @return \Glorand\Model\Settings\Contracts\SettingsManagerContract
+     */
     public function set(string $path, $value): SettingsManagerContract
     {
         $settings = $this->all();