+++ src/Traits/HasSettings.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Glorand\Model\Settings\Traits;
+
+trait HasSettings
+{
+    /**
+     * @return array
+     */
+    abstract public function getSettingsValue(): array;
+}