+++ tests/FieldSettingsManagerTest.php
@@ -1,0 +1,124 @@
+<?php
+
+namespace Glorand\Model\Settings\Tests;
+
+use Glorand\Model\Settings\Tests\Models\UserWithField as User;
+use Glorand\Model\Settings\Tests\Models\WrongUser;
+use Glorand\Model\Settings\Traits\HasSettingsField;
+
+class FieldSettingsManagerTest extends TestCase
+{
+    /** @var \Glorand\Model\Settings\Tests\Models\UserWithField */
+    protected $model;
+    /** @var array */
+    protected $testArray = [
+        'user' => [
+            'first_name' => "John",
+            'last_name'  => "Doe",
+        ],
+    ];
+
+    public function setUp()
+    {
+        parent::setUp();
+        $this->model = User::first();
+    }
+
+    public function testInit()
+    {
+        $traits = class_uses($this->model);
+        $this->assertTrue(array_key_exists(HasSettingsField::class, $traits));
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function testAll()
+    {
+        $this->assertEquals($this->model->settings()->all(), []);
+    }
+
+    /**
+     * @expectedException \Glorand\Model\Settings\Exceptions\ModelSettingsException
+     * @expectedExceptionMessage Unknown field
+     */
+    public function testSettingsMissingSettingsField()
+    {
+        $this->model->settingsFieldName = 'test';
+        $this->model->settings()->all();
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function testHas()
+    {
+        $this->model->settings()->apply($this->testArray);
+        $this->assertEquals($this->model->settings()->all(), $this->testArray);
+
+        $this->assertTrue($this->model->settings()->has('user.first_name'));
+        $this->assertFalse($this->model->settings()->has('user.age'));
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function testGet()
+    {
+        $this->assertEquals($this->model->settings()->all(), []);
+        $this->assertEquals($this->model->settings()->get('user'), null);
+        $this->model->settings()->apply($this->testArray);
+        $this->assertEquals($this->model->settings()->get('user.first_name'), 'John');
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function testApply()
+    {
+        $this->model->settings()->apply($this->testArray);
+        $this->assertEquals($this->model->settings()->all(), $this->testArray);
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function testDelete()
+    {
+        $this->model->settings()->apply($this->testArray);
+
+        $this->assertEquals($this->model->settings()->all(), $this->testArray);
+        $this->assertEquals($this->model->settings()->get('user.first_name'), 'John');
+
+        $this->model->settings()->delete('user.first_name');
+        $this->assertEquals($this->model->settings()->get('user.first_name'), null);
+
+        $this->model->settings()->delete();
+        $this->assertEquals($this->model->settings()->all(), []);
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function testSet()
+    {
+        $this->assertEquals($this->model->settings()->all(), []);
+
+        $this->model->settings()->set('user.age', 18);
+        $this->assertEquals($this->model->settings()->all(), ['user' => ['age' => 18]]);
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function testUpdate()
+    {
+        $this->assertEquals($this->model->settings()->all(), []);
+
+        $this->model->settings()->set('user.age', 18);
+        $this->assertEquals($this->model->settings()->all(), ['user' => ['age' => 18]]);
+
+        $this->model->settings()->update('user.age', 19);
+        $this->assertEquals($this->model->settings()->all(), ['user' => ['age' => 19]]);
+    }
+}