+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 All notable changes to `glorand/laravel-model-settings` will be documented in this file
+
+## 3.0.0 - 2019-01-30
+### Added
+- Command to create the table for settings
+- Dynamic name for the settings table and field
+
 ## 2.0.0 - 2019-01-26
 ### Added
 - Console Commands